requirement.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
websockets==12.0
faster-whisper
piper-tts
soundfile
ffmpeg-python
sentence-transformers
chromadb
faiss-cpu
pypdf
langchain
langchain-community
pydantic
sympy